Move TemplatesListComponent back to TemplatesModule and import from DashboardModule
